Rename local multiplayer directory, consistency in pages
