+++ src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
 
 import Homepage from "./pages/Homepage/Homepage";
-import LocalMultiplayerPage from "./pages/LocalMultiplayerPage/LocalMultiplayerPage";
+import LocalMultiplayerPage from "./pages/LocalMultiplayer/LocalMultiplayerPage";
 import CreateGamePage from "./pages/CreateGame/CreateGamePage";
 
 function App() {
